terraform/requirements.txt
db-dtypes
streamlit 
google-cloud-bigquery 
pandas
